--- src/components/common/textarea.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '@/lib/utils';
-
-function Textarea({ className, ...props }: React.ComponentProps<'textarea'>) {
-  return (
-    <textarea
-      data-slot="textarea"
-      className={cn(
-        'w-full p-[13px] min-h-[86px] bg-navy-opacity-100 border-solid border-[1px] border-border-white rounded-[8px] placeholder:text-text-muted text-[14px] text-text-white resize-none outline-none transition-colors focus-visible:border-ring focus-visible:ring-2 focus-visible:ring-ring/50 disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50 aria-invalid:border-destructive aria-invalid:ring-destructive/20',
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export default Textarea;
